Refactor fluid mutating handler
