+++ pkg/webhook/handler/mutating/mutating_handler_test.go
@@ -813,11 +813,11 @@ func TestAddScheduleInfoToPod(t *testing.T) {
 		fakeClient := fake.NewFakeClientWithScheme(s, objs...)
 		_ = plugins.RegisterMutatingHandlers(fakeClient)
 
-		handler := &CreateUpdatePodForSchedulingHandler{
+		handler := &FluidMutatingHandler{
 			Client: fakeClient,
 		}
 
-		err := handler.AddScheduleInfoToPod(testcase.in, testcase.in.Namespace)
+		err := handler.MutatePod(testcase.in, testcase.in.Namespace)
 		if !((err != nil) == testcase.wantErr) {
 			t.Errorf("testcase %s is failed due to error %v", testcase.name, err)
 		}
@@ -905,7 +905,7 @@ func TestHandle(t *testing.T) {
 	_ = plugins.RegisterMutatingHandlers(fakeClient)
 
 	for _, test := range tests {
-		handler := &CreateUpdatePodForSchedulingHandler{
+		handler := &FluidMutatingHandler{
 			decoder: decoder,
 		}
 		handler.Setup(fakeClient)
@@ -1354,11 +1354,11 @@ func TestAddScheduleInfoToPodWithReferencedDataset(t *testing.T) {
 		fakeClient := fake.NewFakeClientWithScheme(s, objs...)
 		_ = plugins.RegisterMutatingHandlers(fakeClient)
 
-		handler := &CreateUpdatePodForSchedulingHandler{
+		handler := &FluidMutatingHandler{
 			Client: fakeClient,
 		}
 
-		err := handler.AddScheduleInfoToPod(testcase.in, testcase.in.Namespace)
+		err := handler.MutatePod(testcase.in, testcase.in.Namespace)
 		if testcase.wantErr {
 			if err == nil {
 				t.Errorf("testcase %s want error but get nil", testcase.name)
